+++ Helpers/TitleBarHelper.cs
@@ -1,9 +1,7 @@
-﻿using System.Runtime.InteropServices;
-
-using Microsoft.UI;
+﻿using Microsoft.UI;
 using Microsoft.UI.Windowing;
 using Microsoft.UI.Xaml;
-
+using System.Runtime.InteropServices;
 using Windows.UI;
 using Windows.UI.ViewManagement;