+++ Views/OptimizeSystemPage.xaml.cs
@@ -1,12 +1,12 @@
-﻿using System.Diagnostics;
-using System.Text.RegularExpressions;
-using CommunityToolkit.WinUI.Controls;
+﻿using CommunityToolkit.WinUI.Controls;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Media;
 using Microsoft.UI.Xaml.Navigation;
 using Microsoft.Win32;
 using RyTuneX.Helpers;
+using System.Diagnostics;
+using System.Text.RegularExpressions;
 
 namespace RyTuneX.Views;
 
@@ -325,7 +325,7 @@ public sealed partial class OptimizeSystemPage : Page
         try
         {
             var toggleSwitch = (ToggleSwitch)sender;
-            Debug.WriteLine($"ToggleSwitch Tag: {toggleSwitch.Tag}, IsOn: {toggleSwitch.IsOn}");
+            _ = LogHelper.Log($"ToggleSwitch Tag: {toggleSwitch.Tag}, IsOn: {toggleSwitch.IsOn}");
             await OptimizationOptions.XamlSwitchesAsync(toggleSwitch);
         }
         catch (Exception ex)
@@ -410,6 +410,7 @@ public sealed partial class OptimizeSystemPage : Page
         // Handle the Compress button click by setting UI elements and running the command
         compressDialog.PrimaryButtonClick += async (sender, args) =>
         {
+            _ = LogHelper.Log("Starting OS compression");
             CompressOSButton.Visibility = Visibility.Collapsed;
             CompressOSProgressRing.Visibility = Visibility.Visible;
             CompressOSProgressText.Text = "Compressing".GetLocalized();
@@ -423,6 +424,7 @@ public sealed partial class OptimizeSystemPage : Page
         // Handle the Decompress button click by setting UI elements and running the command
         compressDialog.SecondaryButtonClick += async (sender, args) =>
         {
+            _ = LogHelper.Log("Starting OS decompression");
             CompressOSButton.Visibility = Visibility.Collapsed;
             CompressOSProgressRing.Visibility = Visibility.Visible;
             CompressOSProgressText.Text = "Decompressing".GetLocalized();