+++ Views/PrivacyPage.xaml.cs
@@ -1,4 +1,3 @@
-﻿using System.Diagnostics;
 using CommunityToolkit.WinUI.Controls;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
@@ -106,7 +105,7 @@ public sealed partial class PrivacyPage : Page
         try
         {
             var toggleSwitch = (ToggleSwitch)sender;
-            Debug.WriteLine($"ToggleSwitch Tag: {toggleSwitch.Tag}, IsOn: {toggleSwitch.IsOn}");
+            _ = LogHelper.Log($"ToggleSwitch Tag: {toggleSwitch.Tag}, IsOn: {toggleSwitch.IsOn}");
             await OptimizationOptions.XamlSwitchesAsync(toggleSwitch);
         }
         catch (Exception ex)